SpecialFoodTimer done. Next step, ScoreBoard

+++ src/Board.java
@@ -57,6 +57,7 @@ public class Board extends javax.swing.JPanel {
     private int foodDeltaTime;
     private Node next;
     private Node[][] playBoard;
+    private boolean specialFoodVisible;
 
     public Board() {
 
@@ -69,13 +70,18 @@ public class Board extends javax.swing.JPanel {
             public void actionPerformed(ActionEvent ae) {
                 next = createNextNode();
                 gameOver();
-                if (snake.canMove(next.getRow(), next.getCol())) {
-                    if (colideFood()) {
+                //if (snake.canMove(next.getRow(), next.getCol())) {
+                if (colideFood()) {
+                    if (colideNormalFood()) {
                         snake.setRemainingNodesToCreate(1);
-                        food = new Food(snake);
+                        food = new Food(snake, false);
+                    } else {
+                        snake.setRemainingNodesToCreate(4);
+                        specialFood.delete();
                     }
-                    snake.move();
                 }
+                snake.move();
+                //}
                 repaint();
             }
         });
@@ -83,17 +89,17 @@ public class Board extends javax.swing.JPanel {
         specialFoodTimer = new Timer(foodDeltaTime, new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent ae) {
-                specialFood = new Food(snake, true);
-                int random = (int) (Math.random() * 5000 + 7000);
-                Thread threat = new Thread(new Runnable() {
-                    @Override
-                    public void run() {
-                        
-                    }
-                });
-                specialFood.delete();
-                foodDeltaTime = (int) (Math.random() * 10000 + 30000);
-                specialFoodTimer.setDelay(foodDeltaTime);
+                if (!specialFoodVisible) {
+                    specialFood = new Food(snake, true);
+                    int invisibleFoodDeltaTime = (int) (Math.random() * 10000 + 30000);
+                    specialFoodTimer.setDelay(invisibleFoodDeltaTime);
+                    specialFoodVisible = true;
+                } else {
+                    specialFood.delete();
+                    int visibleFoodDeltaTime = (int) (Math.random() * 5000 + 7000);
+                    specialFoodTimer.setDelay(visibleFoodDeltaTime);
+                    specialFoodVisible = false;
+                }
                 repaint();
             }
         });
@@ -106,10 +112,10 @@ public class Board extends javax.swing.JPanel {
 
     private void myInit() {
         snake = new Snake(24, 24, 4);
-        deltaTime = 500;
+        deltaTime = 200;
         foodDeltaTime = 15000;
-        food = new Food(snake);
-        specialFood = new Food(snake, true);
+        food = new Food(snake, false);
+        specialFoodVisible = false;
     }
 
     public Board(int numRows, int numCols) {
@@ -119,18 +125,30 @@ public class Board extends javax.swing.JPanel {
         playBoard = new Node[numRows][numCols];
     }
 
-    public boolean colideFood() {
+    private boolean colideFood() {
+        return colideNormalFood() || colideSpecialFood();
+    }
+
+    private boolean colideNormalFood() {
         return food.getPosition().getRow() == next.getRow() && food.getPosition().getCol() == next.getCol();
+    }
+
+    private boolean colideSpecialFood() {
+        if (specialFoodVisible) {
+            return specialFood.getPosition().getRow() == next.getRow() && specialFood.getPosition().getCol() == next.getCol();
+        }
+        return false;
     }
 
     public void gameOver() {
         if (colideBorders() || colideBody()) {
             snakeTimer.stop();
+            specialFoodTimer.stop();
         }
     }
 
     private boolean colideBorders() {
-        return next.getRow() <= 0 || next.getRow() >= numRows || next.getCol() <= 0 || next.getCol() >= numCols;
+        return next.getRow() < 0 || next.getRow() > numRows || next.getCol() < 0 || next.getCol() > numCols;
     }
 
     private boolean colideBody() {
@@ -174,6 +192,9 @@ public class Board extends javax.swing.JPanel {
         //paintPlayBoard(g2d);
         snake.paint(g2d, squareWidth(), squareHeight());
         food.paint(g2d, squareWidth(), squareHeight());
+        if (specialFoodVisible) {
+            specialFood.paint(g2d, squareWidth(), squareHeight());
+        }
     }
 
     private void paintPlayBoard(Graphics2D g2d) {
